# app/schemas/label.py
from __future__ import annotations

import uuid
from typing import Optional

from pydantic import BaseModel, ConfigDict, Field


class LabelCreate(BaseModel):
    name: str = Field(..., max_length=100)
    description: Optional[str] = Field(default=None, max_length=500)


class LabelResponse(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: uuid.UUID
    scenario_id: uuid.UUID
    name: str
    description: Optional[str]


class MeasurementLabelCreate(BaseModel):
    label_id: uuid.UUID
    reason: Optional[str] = Field(default=None, max_length=500)