# openapi.yaml
openapi: 3.0.3
info:
  title: Environmental Monitoring API
  version: 1.0.0
  description: |
    Production-grade REST API for environmental monitoring, mission management,
    visitor tracking, and geospatial analytics.

    Scenario-scoped routes are prefixed with:
    /api/v1/{scenario_name}/...
servers:
  - url: http://localhost:8000
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/{scenario_name}/measurements/{measurement_id}/labels:
    post:
      tags: [Measurements, Labels]
      summary: Assign label to measurement
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/MeasurementId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementLabelCreate'
      responses:
        '201':
          description: Label assignment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  measurement_id:
                    type: string
                    format: uuid
                  label_id:
                    type: string
                    format: uuid
                  reason:
                    type: string
                    nullable: true
                  created_at:
                    type: string
                    format: date-time
  /api/v1/{scenario_name}/labels:
    get:
      tags: [Labels]
      summary: List labels
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
      responses:
        '200':
          description: Labels list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LabelResponse'
    post:
      tags: [Labels]
      summary: Create label
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelCreate'
      responses:
        '201':
          description: Label created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelResponse'
...
  /api/v1/{scenario_name}/aggregation/map-data:
    get:
      tags: [Aggregation]
      summary: Map feature data for measurements
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - name: lat_min
          in: query
          schema: { type: number, format: double }
        - name: lat_max
          in: query
          schema: { type: number, format: double }
        - name: lon_min
          in: query
          schema: { type: number, format: double }
        - name: lon_max
          in: query
          schema: { type: number, format: double }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 2000, default: 500 }
      responses:
        '200':
          description: Feature collection
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: FeatureCollection
                  features:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
  /api/v1/{scenario_name}/aggregation/measurements/aggregate:
    get:
      tags: [Aggregation]
      summary: Aggregate measurement statistics
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - name: measurement_type_id
          in: query
          schema: { type: string, format: uuid }
        - name: from_date
          in: query
          schema: { type: string, format: date-time }
        - name: to_date
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Aggregation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  aggregations:
                    type: array
                    items:
                      type: object
  /api/v1/{scenario_name}/aggregation/visitors/density:
    get:
      tags: [Aggregation]
      summary: Visitor density grid
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - name: lat_min
          in: query
          schema: { type: number, format: double }
        - name: lat_max
          in: query
          schema: { type: number, format: double }
        - name: lon_min
          in: query
          schema: { type: number, format: double }
        - name: lon_max
          in: query
          schema: { type: number, format: double }
        - name: cell_size
          in: query
          schema: { type: number, format: double, default: 0.01 }
      responses:
        '200':
          description: Density grid
          content:
            application/json:
              schema:
                type: object
                properties:
                  density_grid:
                    type: array
                    items:
                      type: object
                  cell_size_degrees:
                    type: number
components:
  parameters:
    ScenarioName:
      name: scenario_name
      in: path
      required: true
      schema:
        type: string
      description: Scenario identifier/name, e.g. drill-2026
    MeasurementId:
      name: measurement_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LabelCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 100 }
        description:
          type: string
          nullable: true
          maxLength: 500
    LabelResponse:
      type: object
      required: [id, scenario_id, name]
      properties:
        id: { type: string, format: uuid }
        scenario_id: { type: string, format: uuid }
        name: { type: string }
        description:
          type: string
          nullable: true
    MeasurementLabelCreate:
      type: object
      required: [label_id]
      properties:
        label_id:
          type: string
          format: uuid
        reason:
          type: string
          nullable: true
          maxLength: 500
...